Rename `pkg/mock` to `third_party/mock` Rename `component/machinecontrollermanager` to `component/nodemanagement/machinecontrollermanager`
